Change - init tool wont re-create empty repository on DEBUG mode

+++ docker/tools/init.sh
@@ -6,14 +6,13 @@ if [ -z $1 ]; then
 fi
 
 if [ ! -f /.dockerenv ]; then
-    docker compose run app init.sh $1 --rm;
-else
-    echo "Please run this script outside of docker";
-    exit 2;
+    docker compose build;
+    docker compose run app init.sh $1 --rm ;
+    exit;
 fi
 
 # inside docker where in <root>:/app
-if [ ! -d src ]; then
+if [ ! -d src ] && [ -z DEBUG ]; then
     buffalo new $1 --vcs none --skip-docker;
     mv $1 src;
     rm -rf .git;
